Move swapReferences to service
